hid4java already does buffer manip
The hid4java API returns lengths that include both the payload and the extra byte for the report ID. This extra byte can essentially be ignored. Right now this extra byte is special cased in our code to run on Windows - we'll need to test on other OSes to see if it is necessary there as well.